fixed some compiler warnings
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/535



git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@2158 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/unicode/dtfmtsym.h
@@ -399,7 +399,7 @@ private:
 inline void
 DateFormatSymbols::setIsOwned(int32_t which, UBool isOwned)
 {
-    fIsOwned = ( fIsOwned & ~(1 << which) ) | ( (isOwned ? 1 : 0) << which );
+    fIsOwned = (uint8_t)(( fIsOwned & ~(1 << which) ) | ( (isOwned ? 1 : 0) << which ));
 }
 
 inline UBool
